examples: move the v2 examples to the top level. retain the v1 template example for reference but not the others.
