+++ src/c/examples/CMakeLists.txt
@@ -6,7 +6,7 @@ add_subdirectory (discovery)
 if (CURSES_HAVE_CURSES_H)
   add_subdirectory (terminal)
 endif()
-add_subdirectory (v2)
+add_subdirectory (v1)
 add_executable (template template.c)
 target_include_directories (template PRIVATE ../../../include)
 target_link_libraries (template PRIVATE csdk)
