+++ src/c/examples/terminal/device-terminal.c
@@ -1,7 +1,7 @@
 /* Pseudo-device service allowing display of messages in a term using C SDK */
 
 /*
- * Copyright (c) 2018-2019
+ * Copyright (c) 2018-2020
  * IoTech Ltd
  *
  * SPDX-License-Identifier: Apache-2.0
@@ -13,9 +13,13 @@
 #include <curses.h>
 #include <stdio.h>
 
-#include "edgex/devsdk.h"
+#include "devsdk/devsdk.h"
 
-#define ERR_CHECK(x) if (x.code) { fprintf (stderr, "Error: %d: %s\n", x.code, x.reason); edgex_device_service_free (service); free (impl); return x.code; }
+#define ERR_CHECK(x) if (x.code) { fprintf (stderr, "Error: %d: %s\n", x.code, x.reason); devsdk_service_free (service); free (impl); return x.code; }
+#define ERR_BUFSZ 1024
+#define ERR_TERMINAL_READ "GET called for terminal device. This is a write-only device."
+#define ERR_TERMINAL_NO_CMD "No command specified in PUT request."
+#define ERR_TERMINAL_MSG "WriteMsg request did not specify a message."
 
 typedef struct terminal_driver
 {
@@ -24,32 +28,32 @@ typedef struct terminal_driver
 
 static bool terminal_writeMsg
 (
-  terminal_driver *driver,
+  terminal_driver * driver,
   uint32_t nvalues,
-  const edgex_device_commandrequest *requests,
-  const edgex_device_commandresult *values
+  const devsdk_commandrequest * requests,
+  const iot_data_t * values[]
 )
 {
   int x = 0;
   int y = 0;
-  const char *msg = NULL;
+  const char * msg = NULL;
   for (uint32_t i = 0; i < nvalues; i++)
   {
-    const char *param = edgex_nvpairs_value (requests[i].attributes, "parameter");
+    const char * param = devsdk_nvpairs_value (requests[i].attributes, "parameter");
 
     if (param)
     {
-      if (strcmp (param,  "x") == 0)
+      if (strcmp (param, "x") == 0)
       {
-        x = values[i].value.i32_result % COLS;
+        x = (iot_data_i32 (values[i])) % COLS;
       }
       else if (strcmp (param, "y") == 0)
       {
-        y = values[i].value.i32_result % LINES;
+        y = (iot_data_i32 (values[i])) % LINES;
       }
       else if (strcmp (param, "msg") == 0)
       {
-        msg = values[i].value.string_result;
+        msg = iot_data_string (values[i]);
       }
       else if (strcmp (param, "cmd") == 0)
       {
@@ -68,7 +72,6 @@ static bool terminal_writeMsg
 
   if (msg == NULL)
   {
-    iot_log_error (driver->lc, "WriteMsg request did not specify a message");
     return false;
   }
 
@@ -80,9 +83,9 @@ static bool terminal_writeMsg
 }
 
 static bool terminal_init
-  (void *impl, struct iot_logger_t *lc, const edgex_nvpairs *config)
+  (void * impl, struct iot_logger_t * lc, const iot_data_t * config)
 {
-  terminal_driver *driver = (terminal_driver *) impl;
+  terminal_driver * driver = (terminal_driver *) impl;
   driver->lc = lc;
 
   initscr ();
@@ -95,85 +98,108 @@ static bool terminal_init
 
 static bool terminal_get_handler
 (
-  void *impl,
-  const char *devname,
-  const edgex_protocols *protocols,
+  void * impl,
+  const char * devname,
+  const devsdk_protocols * protocols,
   uint32_t nreadings,
-  const edgex_device_commandrequest *requests,
-  edgex_device_commandresult *readings
+  const devsdk_commandrequest * requests,
+  devsdk_commandresult * readings,
+  const devsdk_nvpairs * q_params,
+  iot_data_t ** exception
 )
 {
-  terminal_driver *driver = (terminal_driver *)impl;
-  iot_log_error (driver->lc, "GET called for terminal device. This is a write-only device");
+  terminal_driver * driver = (terminal_driver *) impl;
+  iot_log_error (driver->lc, ERR_TERMINAL_READ);
+  * exception = iot_data_alloc_string (ERR_TERMINAL_READ, IOT_DATA_REF);
   return false;
 }
 
 static bool terminal_put_handler
 (
-  void *impl,
-  const char *devname,
-  const edgex_protocols *protocols,
+  void * impl,
+  const char * devname,
+  const devsdk_protocols * protocols,
   uint32_t nvalues,
-  const edgex_device_commandrequest *requests,
-  const edgex_device_commandresult *values
+  const devsdk_commandrequest * requests,
+  const iot_data_t * values[],
+  iot_data_t ** exception
 )
 {
-  terminal_driver *driver = (terminal_driver *)impl;
-  const char *command = NULL;
+  bool terminal_msg_state = false;
+  terminal_driver * driver = (terminal_driver *) impl;
+  char * buff;
+  const char * command = NULL;
+
   for (uint32_t i = 0; i < nvalues; i++)
   {
-    const char *param = edgex_nvpairs_value (requests[i].attributes, "parameter");
+    const char * param = devsdk_nvpairs_value (requests[i].attributes, "parameter");
     if (param && strcmp (param, "cmd") == 0)
     {
-      command = values[i].value.string_result;
+      command = iot_data_string (values[i]);
       break;
     }
   }
 
   if (command == NULL)
   {
-    iot_log_error (driver->lc, "No command specified in PUT request");
+    iot_log_error (driver->lc, ERR_TERMINAL_NO_CMD);
+    * exception = iot_data_alloc_string (ERR_TERMINAL_NO_CMD, IOT_DATA_REF);
     return false;
   }
   else if (strcmp (command, "WriteMsg") == 0)
   {
-    return terminal_writeMsg (driver, nvalues, requests, values);
+    terminal_msg_state = terminal_writeMsg (driver, nvalues, requests, values);
+    if (!terminal_msg_state)
+    {
+      iot_log_error (driver->lc, ERR_TERMINAL_MSG);
+      * exception = iot_data_alloc_string (ERR_TERMINAL_MSG, IOT_DATA_REF);
+      return false;
+    }
+    else
+    {
+      return true;
+    }
   }
   else
   {
-    iot_log_error (driver->lc, "Unknown command %s", command);
+    buff = malloc (ERR_BUFSZ);
+    snprintf (buff, ERR_BUFSZ, "Unknown command %s", command);
+    iot_log_error (driver->lc, buff);
+    * exception = iot_data_alloc_string (buff, IOT_DATA_TAKE);
     return false;
   }
 }
 
-/* ---- Disconnect ---- */
-/* Disconnect handles protocol-specific cleanup when a device is removed. */
-static bool terminal_disconnect (void *impl, edgex_protocols *device)
-{
-  return true;
-}
-
 /* ---- Stop ---- */
 /* Stop performs any final actions before the device service is terminated */
-static void terminal_stop (void *impl, bool force)
+static void terminal_stop (void * impl, bool force)
 {
   endwin ();
 }
 
-
-int main (int argc, char *argv[])
+int main (int argc, char * argv[])
 {
-  edgex_device_svcparams params = { "device-terminal", NULL, NULL, NULL };
   sigset_t set;
   int sigret;
 
   terminal_driver * impl = malloc (sizeof (terminal_driver));
   impl->lc = NULL;
 
-  if (!edgex_device_service_processparams (&argc, argv, &params))
+  devsdk_error e;
+  e.code = 0;
+
+  devsdk_callbacks terminalImpls =
   {
-    return  0;
-  }
+    terminal_init,         /* Initialize */
+    NULL,                 /* Discovery */
+    terminal_get_handler,  /* Get */
+    terminal_put_handler,  /* Put */
+    terminal_stop          /* Stop */
+  };
+
+  devsdk_service_t * service = devsdk_service_new
+    ("device-terminal", "1.0", impl, terminalImpls, &argc, argv, &e);
+  ERR_CHECK (e);
 
   int n = 1;
   while (n < argc)
@@ -181,8 +207,7 @@ int main (int argc, char *argv[])
     if (strcmp (argv[n], "-h") == 0 || strcmp (argv[n], "--help") == 0)
     {
       printf ("Options:\n");
-      printf ("  -h, --help\t\t: Show this text\n");
-      edgex_device_service_usage ();
+      printf ("  -h, --help\t\t\tShow this text\n");
       return 0;
     }
     else
@@ -192,34 +217,17 @@ int main (int argc, char *argv[])
     }
   }
 
-  edgex_error e;
-  e.code = 0;
-
-  edgex_device_callbacks terminalImpls =
-  {
-    terminal_init,         /* Initialize */
-    NULL,                 /* Discovery */
-    terminal_get_handler,  /* Get */
-    terminal_put_handler,  /* Put */
-    terminal_disconnect,   /* Disconnect */
-    terminal_stop          /* Stop */
-  };
-
-  edgex_device_service *service = edgex_device_service_new
-    (params.svcname, "1.0", impl, terminalImpls, &e);
-  ERR_CHECK (e);
-
-  edgex_device_service_start (service, params.regURL, params.profile, params.confdir, &e);
+  devsdk_service_start (service, &e);
   ERR_CHECK (e);
 
   sigemptyset (&set);
   sigaddset (&set, SIGINT);
   sigwait (&set, &sigret);
 
-  edgex_device_service_stop (service, true, &e);
+  devsdk_service_stop (service, true, &e);
   ERR_CHECK (e);
 
-  edgex_device_service_free (service);
+  devsdk_service_free (service);
   free (impl);
   return 0;
 }
