+++ src/c/examples/v1/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_executable (templatev1 template.c)
+target_include_directories (templatev1 PRIVATE ../../../../include)
+target_link_libraries (templatev1 PRIVATE csdk)
