+++ src/c/examples/v1/README.md
@@ -1,0 +1,7 @@
+## Template device service (v1 API)
+
+### About
+
+This is a version of the template device service written for v1 of the CSDK API.
+It is provided for information only; use of the v2 API is recommended for any
+new device service.
