--- src/c/examples/v2/CMakeLists.txt
@@ -1,10 +1,0 @@
-find_package (Curses)
-add_subdirectory (counters)
-add_subdirectory (gyro)
-add_subdirectory (random)
-if (CURSES_HAVE_CURSES_H)
-    add_subdirectory (terminal)
-endif()
-add_executable (template2 template.c)
-target_include_directories (template2 PRIVATE ../../../../include)
-target_link_libraries (template2 PRIVATE csdk)
