--- src/c/examples/v2/counters/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_executable (device-counterv2 device-counter.c)
-target_include_directories (device-counterv2 PRIVATE ../../../../../include)
-target_link_libraries (device-counterv2 PRIVATE csdk)
