--- src/c/examples/v2/counters/res/configuration.toml
@@ -1,57 +1,0 @@
-[Service]
-  Port = 49999
-  Timeout = 5000
-  ConnectRetries = 3
-  Labels = [ "Counter" ]
-  StartupMsg = "Example counting device started"
-  CheckInterval = "10s"
-
-[Clients]
-  [Clients.Data]
-    Host = "localhost"
-    Port = 48080
-
-  [Clients.Metadata]
-    Host = "localhost"
-    Port = 48081
-
-  [Clients.Logging]
-    Host = "localhost"
-    Port = 48061
-
-[Device]
-  DataTransform = false
-  Discovery = false
-  InitCmd = ""
-  InitCmdArgs = ""
-  MaxCmdOps = 128
-  MaxCmdResultLen = 256
-  RemoveCmd = ""
-  RemoveCmdArgs = ""
-  ProfilesDir = ""
-  SendReadingsOnChanged = true
-
-[Logging]
-  EnableRemote = false
-  File = "-"
-  LogLevel = "DEBUG"
-
-[[DeviceList]]
-  Name = "Counter1"
-  Profile = "Example-Counter"
-  Description = "An Example Counting Device"
-  [DeviceList.Protocols]
-    [DeviceList.Protocols.Counter]
-      Index = "3"
-  [[DeviceList.AutoEvents]]
-    Resource = "Counter"
-    OnChange = false
-    Frequency = "10s"
-
-[[DeviceList]]
-  Name = "Counter2"
-  Profile = "Example-Counter"
-  Description = "Another Example Counting Device"
-  [DeviceList.Protocols]
-    [DeviceList.Protocols.Counter]
-      Index = "58"
