--- src/c/examples/v2/gyro/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_executable (device-gyrov2 device-gyro.c)
-target_include_directories (device-gyrov2 PRIVATE ../../../../../include)
-target_link_libraries (device-gyrov2 PRIVATE csdk)
