--- src/c/examples/v2/gyro/res/Gyro.yaml
@@ -1,60 +1,0 @@
-name: "Example-Gyro"
-manufacturer: "IoTechSystems"
-model: "IoT4"
-description: "Virtual 3-axis Gyro Device for CSDK Example"
-labels:
-  - "sensor"
-
-deviceResources:
-  -
-    name: Xrotation
-    description: "X axis rotation rate"
-    attributes:
-      { parameter: "xrot" }
-    properties:
-      value:
-        { type: "int32", readWrite: "R" }
-      units:
-        { defaultValue: "degrees/sec" }
-  -
-    name: Yrotation
-    description: "Y axis rotation rate"
-    attributes:
-      { parameter: "yrot" }
-    properties:
-      value:
-        { type: "int32", readWrite: "R" }
-      units:
-        { defaultValue: "degrees/sec" }
-  -
-    name: Zrotation
-    description: "Z axis rotation rate"
-    attributes:
-      { parameter: "zrot" }
-    properties:
-      value:
-        { type: "int32", readWrite: "R" }
-      units:
-        { defaultValue: "degrees/sec" }
-
-deviceCommands:
-  -
-    name: rotation
-    get:
-      - { deviceResource: "Xrotation" }
-      - { deviceResource: "Yrotation" }
-      - { deviceResource: "Zrotation" }
-
-coreCommands:
-  -
-    name: rotation
-    get:
-      path: "/api/v1/device/{deviceId}/rotation"
-      responses:
-        - code: "200"
-          description: "Successfully read the rotation sensors."
-          expectedValues: [ "Xrotation", "Yrotation", "Zrotation" ]
-        - code: "503"
-          description: "service unavailable"
-          expectedValues: []
-
