--- src/c/examples/v2/gyro/res/configuration.toml
@@ -1,44 +1,0 @@
-[Service]
-  Port = 49999
-  Timeout = 5000
-  ConnectRetries = 10
-  Labels = [ 'Gyro' ]
-  StartupMsg = 'Example gyro device service started'
-  CheckInterval = '10s'
-
-[Clients]
-  [Clients.Data]
-    Host = 'localhost'
-    Port = 48080
-
-  [Clients.Metadata]
-    Host = 'localhost'
-    Port = 48081
-
-  [Clients.Logging]
-    Host = 'localhost'
-    Port = 48061
-
-[Device]
-  DataTransform = false
-  Discovery = false
-  InitCmd = ''
-  InitCmdArgs = ''
-  MaxCmdOps = 128
-  MaxCmdResultLen = 256
-  RemoveCmd = ''
-  RemoveCmdArgs = ''
-  ProfilesDir = ''
-  SendReadingsOnChanged = true
-
-[Logging]
-  EnableRemote = false
-  File = '-'
-  LogLevel = 'DEBUG'
-
-[[DeviceList]]
-  Name = 'Gyro'
-  Profile = 'Example-Gyro'
-  Description = 'An Example Gyro Device'
-  [DeviceList.Protocols]
-    [DeviceList.Protocols.Gyro]
