--- src/c/examples/v2/random/res/configuration.toml
@@ -1,53 +1,0 @@
-[Service]
-  Port = 49999
-  Timeout = 5000
-  ConnectRetries = 10
-  Labels = [ 'Random' ]
-  StartupMsg = 'Example random device started'
-  CheckInterval = '10s'
-
-[Clients]
-  [Clients.Data]
-    Host = 'localhost'
-    Port = 48080
-
-  [Clients.Metadata]
-    Host = 'localhost'
-    Port = 48081
-
-  [Clients.Logging]
-    Host = 'localhost'
-    Port = 48061
-
-[Device]
-  DataTransform = true
-  Discovery = false
-  InitCmd = ''
-  InitCmdArgs = ''
-  MaxCmdOps = 128
-  MaxCmdResultLen = 256
-  RemoveCmd = ''
-  RemoveCmdArgs = ''
-  ProfilesDir = ''
-  SendReadingsOnChanged = true
-
-[Logging]
-  EnableRemote = false
-  File = '-'
-  LogLevel = 'DEBUG'
-
-[[DeviceList]]
-  Name = 'RandomDevice1'
-  Profile = 'ExampleSensor'
-  Description = 'An Example Random Device'
-  [DeviceList.Protocols]
-    [DeviceList.Protocols.Other]
-      Address = 'someSensorEndpoint'
-  [[DeviceList.AutoEvents]]
-    Resource = 'SensorOne'
-    OnChange = false
-    Frequency = '10s'
-  [[DeviceList.AutoEvents]]
-    Resource = 'SensorTwo'
-    OnChange = true
-    Frequency = '15000ms'
