--- src/c/examples/v2/terminal/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_executable (device-terminalv2 device-terminal.c)
-target_include_directories (device-terminalv2 PRIVATE ../../../../../include)
-target_link_libraries (device-terminalv2 PRIVATE csdk ncurses)
